Adds tests for admin page
